[#132] Add test_lib module for common test needs
There are several functions used by more than one test suite which
could be pulled into a separate file to reduce code duplication. That
is what this commit does.
